Moved ContentPreparer to posts directory
ref https://linear.app/ghost/issue/AP-961

Now that this is soley used by the Post entity it makes sense for them to live
together as this is now business logic of how posts are created.
